+++ src/components/FollowingListContainer.tsx
@@ -5,7 +5,7 @@ import UserApplicationService from '../application/UserApplicationService';
 import ScreenName from '../domain/models/User/Profile/ScreenName';
 import { UserDataModel } from '../infrastructure/UserDataModel';
 
-// todo ここでデータを取得したら list を作ってくれるコンポーネントにデータを渡す。
+// ここでデータを取得したら list を作ってくれるコンポーネントにデータを渡す。
 export const FollowingListContainer: React.FC = () => {
   const [userIndicating, setUserIndicating] = React.useState();
   const [existUser, setExistUser] = React.useState();
