+++ src/components/SignUpContainer.tsx
@@ -87,9 +87,11 @@ export const SignUpContainer: React.FC = () => {
     return d <= lastDay(m, y);
   };
 
+  // MEMO isValidHoge はあくまでエラーメッセージコンポーネントを表示するためのフラグなのでここで改めて判定する必要がある。
+  // TODO ここでやるべきではない
+  // TODO やる
+  // TODO わかりずらいので説明変数を使おう
   const judgeCanGoNextPage = ({
-    // isValidHoge はあくまでエラーメッセージコンポーネントを表示するためのフラグなのでここで改めて判定する必要がある。
-    // todo ここでやるべきではない
     isRightUserName = userName.length > 0 &&
       userName.length <= MAX_USER_NAME_LENGTH,
     isRightScreenName = screenName.length > 0 &&
